pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "cuvis_ai_core"
version = "0.1.0"
description = "Cuvis-AI Core Framework - Hyperspectral image processing pipeline framework"
authors = [
    {name = "Cubert GmbH", email = "support@cubert.com"}
]
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.10"

dependencies = [
    "cuvis==3.4.5",
    # Core framework dependencies
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.1.0",
    # Configuration and orchestration
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    # gRPC and communication
    "grpcio>=1.56.0",
    "protobuf>=4.23.0",
    "grpcio-health-checking>=1.56.0",
    # Utilities and logging
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    # Plugin system
    "gitpython>=3.1.40",
    # Type hints and validation
    "typing-extensions>=4.0.0",
    "pydantic>=2.12.5",
    "gdown>=5.2.1",
    "dataclass-wizard",
    "dotenv>=0.9.9",
    "ruff>=0.14.11",
]

[project.optional-dependencies]

dev = [
    "ipykernel",
    "jupyterlab",
    "ruff",
    "pytest",
    "pytest-cov",
    "ipympl>=0.9.8",
    "ipdb>=0.13.13",
    "psutil",
]


test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[project.urls]
Homepage = "https://gitlab.cubert.local/cubert/cuvis-ai-core"
Repository = "https://gitlab.cubert.local/cubert/cuvis-ai-core.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["cuvis_ai_core"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
